requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pika==1.3.2
openai==1.3.0
pymongo==4.6.0
python-dotenv==1.0.0
typing-extensions==4.8.0
loguru==0.7.2
